Move MySQL test to its own module
