+++ introspection-engine/introspection-engine-tests/tests/relations/mysql.rs
@@ -1,6 +1,8 @@
 use barrel::types;
 use expect_test::expect;
+use indoc::indoc;
 use introspection_engine_tests::test_api::*;
+use quaint::prelude::Queryable;
 use test_macros::test_connector;
 
 #[test_connector(tags(Mysql))]
@@ -636,6 +638,48 @@ async fn one_to_one_relation_on_a_singular_primary_key(api: &TestApi) -> TestRes
     "#]];
 
     expected.assert_eq(&api.introspect_dml().await?);
+
+    Ok(())
+}
+
+#[test_connector(tags(Mysql57))]
+async fn multiple_foreign_key_constraints_are_taken_always_in_the_same_order(api: &TestApi) -> TestResult {
+    let migration = indoc! {r#"
+        CREATE TABLE A
+        (
+            id  int primary key,
+            foo int not null
+        );
+
+        CREATE TABLE B
+        (
+            id int primary key
+        );
+
+        ALTER TABLE A ADD CONSTRAINT fk_1 FOREIGN KEY (foo) REFERENCES B(id) ON DELETE CASCADE ON UPDATE CASCADE;
+        ALTER TABLE A ADD CONSTRAINT fk_2 FOREIGN KEY (foo) REFERENCES B(id) ON DELETE RESTRICT ON UPDATE RESTRICT;
+    "#};
+
+    api.database().raw_cmd(migration).await?;
+
+    let expected = expect![[r#"
+        model A {
+          id  Int @id
+          foo Int
+          B   B   @relation(fields: [foo], references: [id], onDelete: Cascade)
+
+          @@index([foo], map: "fk_2")
+        }
+
+        model B {
+          id Int @id
+          A  A[]
+        }
+    "#]];
+
+    for _ in 0..10 {
+        expected.assert_eq(&api.introspect_dml().await?);
+    }
 
     Ok(())
 }
